pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "stoik"
version = "0.2.0"
description = "Stream-to-store pipeline framework: Kafka → buffer → batch flush → DuckDB"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = ["duckdb", "pyarrow", "structlog", "confluent-kafka"]

[project.optional-dependencies]
flight = ["flightsql-dbapi", "protobuf"]
all = ["stoik[flight]"]

[project.urls]
Repository = "https://github.com/Kord96/stoik"

[tool.setuptools.packages.find]
include = ["stoik*"]
